# backend/src/api/oauth/yandex.rs
use axum::{
    extract::{Query, State},
    http::HeaderMap,
    response::Redirect,
};
use tower_cookies::Cookies;

use crate::{
    api::oauth::{YandexState, redirect_with_error},
    domain::oauth::service::{OAuthProvider, OAuthService},
    dto::oauth::OAuthCallbackQuery,
    errors::LMSError,
    utils::{add_cookie, device_from_headers, remove_cookie},
};

/// Redirect user to yandex oauth login page
#[utoipa::path(get, path = "/login", tag = "OAuth")]
pub async fn login(cookies: Cookies, State(state): State<YandexState>) -> Redirect {
    let (oauth_state, code_verifier, code_challenge) = OAuthService::generate();

    add_cookie(&cookies, ("oauth_state", oauth_state.clone()));
    add_cookie(&cookies, ("code_verifier", code_verifier));

    let url = state.yandex_provider.url(oauth_state, code_challenge);
    Redirect::temporary(url.as_str())
}

/// Callback for yandex oauth provider
#[utoipa::path(get, path = "/callback", tag = "OAuth")]
pub async fn callback(
    cookies: Cookies,
    headers: HeaderMap,
    Query(query): Query<OAuthCallbackQuery>,
    State(state): State<YandexState>,
) -> Result<Redirect, LMSError> {
    let (oauth_state, code_verifier) = OAuthService::parse_cookies(&cookies)?;

    if oauth_state != query.state {
        return Err(LMSError::Forbidden("Invalid `state` parameter".to_string()));
    }

    let user = state
        .yandex_provider
        .get_user(query.code, code_verifier)
        .await?;

    let user_id = match state.oauth_service.save_user(user).await {
        Err(err @ LMSError::EmailMissing) => {
            return Ok(redirect_with_error(
                &cookies,
                &state.account_service.redirect_url,
                &err,
            ));
        }
        result => result?,
    };

    let (refresh_token, _) = state
        .refresh_token_service
        .create_refresh_token(user_id, device_from_headers(&headers))
        .await?;

    let account = state.account_service.get_user(user_id).await?;
    let access_token = state.jwt.generate_access_token(
        user_id,
        account.role,
        account.email_verified,
        account.is_profile_complete(),
    )?;

    add_cookie(&cookies, ("refresh_token", refresh_token));

    remove_cookie(&cookies, "oauth_state");
    remove_cookie(&cookies, "code_verifier");

    Ok(Redirect::to(
        format!(
            "{}?access_token={access_token}",
            state.account_service.redirect_url
        )
        .as_str(),
    ))
}