# frontend/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "utoipa-axum",
		"description": "Utoipa's axum bindings for seamless integration for the two",
		"contact": {
			"name": "Juha Kukkonen",
			"email": "juha7kukkonen@gmail.com"
		},
		"license": {
			"name": "MIT OR Apache-2.0",
			"identifier": "MIT OR Apache-2.0"
		},
		"version": "0.2.0"
	},
	"paths": {
		"/account": {
			"get": {
				"tags": ["Account"],
				"summary": "Return user object",
				"operationId": "get_user",
				"responses": {
					"200": {
						"description": "",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/GetUserResponseDTO"
								}
							}
						}
					}
				},
				"security": [
					{
						"BearerAuth": []
					}
				]
			}
		},
...
	},
	"components": {
		"schemas": {
...
			"GetUserResponseDTO": {
				"type": "object",
				"required": [
					"id",
					"username",
					"email",
					"role",
					"email_verified",
					"profile_complete"
				],
				"properties": {
					"email": {
						"type": "string"
					},
					"email_verified": {
						"type": "boolean"
					},
					"first_name": {
						"type": ["string", "null"]
					},
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"last_name": {
						"type": ["string", "null"]
					},
					"patronymic": {
						"type": ["string", "null"]
					},
					"profile_complete": {
						"type": "boolean"
					},
					"role": {
						"$ref": "#/components/schemas/UserRole"
					},
					"username": {
						"type": "string"
					}
				}
			},
...
			"UserRole": {
				"type": "string",
				"enum": ["Student", "Teacher", "Admin"]
			},
...
		}
	}
}